add doc comments on options
